chore: Restructure folders
