+++ web/app/(landing)/page.tsx
@@ -1,4 +1,4 @@
-import LandingPage from '@/components/ui/landing/ui'
+import LandingPage from './landing'
 
 export default function Home() {
     return <LandingPage />
