+++ web/app/(landing)/types/index.ts
