+++ web/app/(sandbox)/action.ts
@@ -6,6 +6,7 @@ import type { CodeResponse } from './types'
 import * as Sentry from '@sentry/nextjs'
 
 import { isNonEmptyString } from '@/lib/utils'
+
 import { getImage, RCEHandler } from './core'
 
 export default async function handleCodeExecution({
