+++ web/app/layout.tsx
@@ -1,4 +1,4 @@
-import type { Metadata } from 'next'
+import type { Metadata, Viewport } from 'next'
 import type { ReactNode } from 'react'
 
 import { Analytics } from '@vercel/analytics/react'
@@ -6,9 +6,9 @@ import { SpeedInsights } from '@vercel/speed-insights/next'
 
 import ReduxProvider from '@/components/providers/redux-provider'
 import ThemeProvider from '@/components/providers/theme-provider'
+import { jetbrainsMono, mukta } from '@/config/fonts'
 import siteConfig from '@/config/site'
 import '@/styles/global.css'
-import { mukta, jetbrainsMono } from '@/config/fonts'
 
 export const metadata: Metadata = {
     metadataBase: new URL(siteConfig.siteUrl),
@@ -28,40 +28,25 @@ export const metadata: Metadata = {
             url: '/apple-touch-icon.png',
         },
     ],
-    applicationName: siteConfig.name,
-    authors: [
-        {
-            name: siteConfig.author,
-            url: new URL(siteConfig.socials.authorUrl),
-        },
-    ],
-    generator: siteConfig.publisher,
-    creator: siteConfig.author,
-    publisher: siteConfig.publisher,
     robots: 'index, follow',
     keywords: siteConfig.keywords,
     referrer: 'origin',
     description: siteConfig.description,
     openGraph: {
-        title: siteConfig.title,
-        description: siteConfig.description,
-        siteName: siteConfig.name,
         images: siteConfig.OGImage,
-        locale: siteConfig.locale,
         type: 'website',
+        url: new URL(siteConfig.siteUrl),
     },
     twitter: {
-        title: {
-            default: siteConfig.title,
-            template: `%s | ${siteConfig.title}`,
-        },
         card: 'summary_large_image',
         creator: siteConfig.twitterHandle,
-        description: siteConfig.description,
         images: siteConfig.OGImage,
     },
-    formatDetection: { telephone: false },
     alternates: { canonical: siteConfig.siteUrl },
+}
+
+export const viewport: Viewport = {
+    maximumScale: 1,
 }
 
 export default function RootLayout({ children }: { children: ReactNode }) {
