+++ web/app/robots.ts
@@ -1,16 +1,8 @@
-import { MetadataRoute } from 'next'
+import type { MetadataRoute } from 'next'
 
-import siteConfig from '@/config/site'
-
-const robots = (): MetadataRoute.Robots => {
-    return {
-        rules: {
-            userAgent: '*',
-            allow: '/',
-            disallow: '/api/',
-        },
-        sitemap: `${siteConfig.siteUrl}/sitemap.xml`,
-    }
-}
-
-export default robots
+export default {
+    rules: {
+        userAgent: '*',
+        allow: '/',
+    },
+} satisfies MetadataRoute.Robots
