--- web/app/sitemap.ts
@@ -1,16 +1,0 @@
-import { MetadataRoute } from 'next'
-
-import siteConfig from '@/config/site'
-
-const sitemap = (): MetadataRoute.Sitemap => {
-    return [
-        {
-            url: siteConfig.siteUrl,
-            lastModified: new Date(),
-            changeFrequency: 'weekly',
-            priority: 1.0,
-        },
-    ]
-}
-
-export default sitemap
